A better logic and some fixes
Added a better logic for the reply msgs sent by the bot to the members who ask movies in the support group. (IT MUST WORK NOW). And some fixes for the callbacks of the buttons which is not working for the start msgs on groups which is sent by the bot.